requirements.txt
fastapi
uvicorn[standard]==0.24.0
azure-search-documents==11.6.0b1
python-dotenv==1.1.0
requests==2.31.0
pydantic==2.8.2
pydantic-settings==2.10.1

# New dependencies for GitHub integration
aiofiles==23.2.1
httpx
openai
anyio

# Security and rate limiting
slowapi==0.1.9

# MCP SDK (optional - install with pip install 'mcp[cli]')
# mcp[cli]
fastmcp

# Microsoft Docs MCP integration
aiohttp==3.9.1

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5
